backend/requirements.txt
fastapi==0.110.1
uvicorn==0.25.0
python-dotenv>=1.0.1
pymongo==4.5.0
motor==3.3.1
pydantic[email]>=2.6.4
starlette>=0.37.2
passlib[bcrypt]>=1.7.4
bcrypt==4.1.3
pyjwt>=2.10.0
cryptography>=42.0.8
httpx>=0.27.0
emergentintegrations==0.1.0
